Agregar logica fin uso de instalacion

+++ TP5/Formularios/Form1.cs
@@ -52,7 +52,7 @@ namespace TP5
                 reloj_max = int.Parse(txt_tiempo_simular.Text);
                 consulta_min = double.Parse(txt_consulta_min.Text);
                 consulta_max = double.Parse(txt_consulta_max.Text);
-                porc_retiran_biblo = int.Parse(txt_probabilidad_retirarse.Text) / 100;
+                porc_retiran_biblo = double.Parse(txt_probabilidad_retirarse.Text) / 100;
                 tiempo_uso_instalacion = int.Parse(txt_tiempo_uso_instalacion.Text);
                 prob_pedido = double.Parse(txt_pedido.Text) / 100;
                 prob_devolucion = double.Parse(txt_devolucion.Text) / 100;
